+++ src/Application/Features/Employees/GetUserEmployeesQuery.cs
@@ -22,6 +22,6 @@ public record GetUserEmployeesQueryResponse(
     string? JobTitle,
     EmployeeStatus Status,
     Region Region,
-    decimal? LeaveUsedInCurrentYear = null,
-    decimal? NetWorkedDaysBeforeCurrentMonth = null,
-    decimal? CarriedOverLeaveFromPreviousYear = null);
+    int? LeaveUsedInCurrentYear = null,
+    int? NetWorkedDaysBeforeCurrentMonth = null,
+    int? CarriedOverLeaveFromPreviousYear = null);
