+++ src/Core/Contracts/Employees/EmployeeDto.cs
@@ -11,6 +11,6 @@ public record EmployeeDto(
     string PhoneNumber,
     string? JobTitle,
     Region? Region,
-    decimal? LeaveUsedInCurrentYear = null,
-    decimal? NetWorkedDaysBeforeCurrentMonth = null,
-    decimal? CarriedOverLeaveFromPreviousYear = null);
+    int? LeaveUsedInCurrentYear = null,
+    int? NetWorkedDaysBeforeCurrentMonth = null,
+    int? CarriedOverLeaveFromPreviousYear = null);
