+++ src/Core/Contracts/Employees/UserEmployeeByIdResult.cs
@@ -12,7 +12,7 @@ public record UserEmployeeByIdResult(
     string PhoneNumber,
     string? JobTitle,
     Region Region,
-    decimal? LeaveUsedInCurrentYear,
-    decimal? NetWorkedDaysBeforeCurrentMonth,
-    decimal? CarriedOverLeaveFromPreviousYear,
+    int? LeaveUsedInCurrentYear,
+    int? NetWorkedDaysBeforeCurrentMonth,
+    int? CarriedOverLeaveFromPreviousYear,
     List<EmployeeBankAccountDto> BankAccounts);
