+++ src/Core/Contracts/Employees/UserEmployeeResult.cs
@@ -11,6 +11,6 @@ public record UserEmployeeResult(
     string? JobTitle,
     EmployeeStatus Status,
     Region Region,
-    decimal? LeaveUsedInCurrentYear = null,
-    decimal? NetWorkedDaysBeforeCurrentMonth = null,
-    decimal? CarriedOverLeaveFromPreviousYear = null);
+    int? LeaveUsedInCurrentYear = null,
+    int? NetWorkedDaysBeforeCurrentMonth = null,
+    int? CarriedOverLeaveFromPreviousYear = null);
